Move encryption modules to directory
