+++ src/renderer/modules/encryption/credentials.js
