+++ src/renderer/modules/encryption/salt.js
